Add files via upload
Detailed description: https://www.ers.usda.gov/data-products/food-environment-atlas/data-access-and-documentation-downloads/
The current version of the Food Environment Atlas has over 275 variables, including new indicators on access and proximity to a grocery store for sub populations; an indicator on the SNAP Combined Application Project for recipients of Supplemental Security Income (at the State level); and indicators on farmers' markets that report accepting credit cards or report selling baked and prepared food products. All of the data included in the Atlas are aggregated into an Excel spreadsheet for easy download. These data are from a variety of sources and cover varying years and geographic levels. The documentation for each version of the data provides complete information on definitions and data sources.

In the downloadable Excel spreadsheets:
State and county Federal information processing standards (FIPS) codes are provided.
The variable lookup file links the short field descriptions (indicator names) used in the data file with the longer indicator names used in the Atlas.

Unless otherwise noted with asterisks on the longer indicator names (in the variable lookup file), indicators are county-level measures. A single asterisk * denotes a State-level indicator, while a double asterisk ** denotes a regional-level indicator.
"No data" fields are empty or referenced with "-9999".

Supplemental data are provided in additional tabs (State- and county-level data are provided separately).